pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["dimansidorov <dimkasidorow@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
sqlalchemy = "^2.0.4"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
poetry-dotenv-plugin = "^0.1.0"
bcrypt-flask = "^1.0.2"
wtforms = "^3.0.1"
dnspython = "^2.3.0"
email-validator = "^1.3.1"
flask-wtf = "^1.1.1"
flask-admin = "^1.6.1"
flask-paginate = "^2022.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
